Move classes to the corresponding packages again
