+++ src/rider/main/kotlin/controlmessages/AvaloniaMessages.kt
@@ -1,7 +1,8 @@
-package me.fornever.avaloniarider
+package me.fornever.avaloniarider.controlmessages
 
 import com.jetbrains.rider.util.idea.application
 import com.jetbrains.rider.util.idea.getComponent
+import me.fornever.avaloniarider.toUUID
 import java.util.*
 
 class AvaloniaMessages {
